% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_cancel_deployment_job}
\alias{robomaker_cancel_deployment_job}
\title{Cancels the specified deployment job}
\usage{
robomaker_cancel_deployment_job(job)
}
\arguments{
\item{job}{[required] The deployment job ARN to cancel.}
}
\value{
An empty list.
}
\description{
Cancels the specified deployment job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_deployment_job(
  job = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
