% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_cancel_simulation_job}
\alias{robomaker_cancel_simulation_job}
\title{Cancels the specified simulation job}
\usage{
robomaker_cancel_simulation_job(job)
}
\arguments{
\item{job}{[required] The simulation job ARN to cancel.}
}
\value{
An empty list.
}
\description{
Cancels the specified simulation job.
}
\section{Request syntax}{
\preformatted{svc$cancel_simulation_job(
  job = "string"
)
}
}

\keyword{internal}
