% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_cancel_simulation_job_batch}
\alias{robomaker_cancel_simulation_job_batch}
\title{Cancels a simulation job batch}
\usage{
robomaker_cancel_simulation_job_batch(batch)
}
\arguments{
\item{batch}{[required] The id of the batch to cancel.}
}
\value{
An empty list.
}
\description{
Cancels a simulation job batch. When you cancel a simulation job batch,
you are also cancelling all of the active simulation jobs created as
part of the batch.
}
\section{Request syntax}{
\preformatted{svc$cancel_simulation_job_batch(
  batch = "string"
)
}
}

\keyword{internal}
