% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_cancel_world_export_job}
\alias{robomaker_cancel_world_export_job}
\title{Cancels the specified export job}
\usage{
robomaker_cancel_world_export_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (arn) of the world export job to cancel.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Cancels the specified export job.
}
\section{Request syntax}{
\preformatted{svc$cancel_world_export_job(
  job = "string"
)
}
}

\keyword{internal}
