% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_cancel_world_generation_job}
\alias{robomaker_cancel_world_generation_job}
\title{Cancels the specified world generator job}
\usage{
robomaker_cancel_world_generation_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (arn) of the world generator job to cancel.}
}
\value{
An empty list.
}
\description{
Cancels the specified world generator job.
}
\section{Request syntax}{
\preformatted{svc$cancel_world_generation_job(
  job = "string"
)
}
}

\keyword{internal}
