% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_robot_application_version}
\alias{robomaker_create_robot_application_version}
\title{Creates a version of a robot application}
\usage{
robomaker_create_robot_application_version(application,
  currentRevisionId)
}
\arguments{
\item{application}{[required] The application information for the robot application.}

\item{currentRevisionId}{The current revision id for the robot application. If you provide a
value and it matches the latest revision ID, a new version will be
created.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  name = "string",
  version = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      etag = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  robotSoftwareSuite = list(
    name = "ROS"|"ROS2",
    version = "Kinetic"|"Melodic"|"Dashing"
  ),
  lastUpdatedAt = as.POSIXct(
    "2015-01-01"
  ),
  revisionId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a version of a robot application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_robot_application_version(
  application = "string",
  currentRevisionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
