% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_world_generation_job}
\alias{robomaker_create_world_generation_job}
\title{Creates worlds using the specified template}
\usage{
robomaker_create_world_generation_job(clientRequestToken, template,
  worldCount, tags, worldTags)
}
\arguments{
\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{template}{[required] The Amazon Resource Name (arn) of the world template describing the
worlds you want to create.}

\item{worldCount}{[required] Information about the world count.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
world generator job.}

\item{worldTags}{A map that contains tag keys and tag values that are attached to the
generated worlds.}
}
\value{
A list with the following syntax:\preformatted{list(
  arn = "string",
  status = "Pending"|"Running"|"Completed"|"Failed"|"PartialFailed"|"Canceling"|"Canceled",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  failureCode = "InternalServiceError"|"LimitExceeded"|"ResourceNotFound"|"RequestThrottled"|"InvalidInput"|"AllWorldGenerationFailed",
  clientRequestToken = "string",
  template = "string",
  worldCount = list(
    floorplanCount = 123,
    interiorCountPerFloorplan = 123
  ),
  tags = list(
    "string"
  ),
  worldTags = list(
    "string"
  )
)
}
}
\description{
Creates worlds using the specified template.
}
\section{Request syntax}{
\preformatted{svc$create_world_generation_job(
  clientRequestToken = "string",
  template = "string",
  worldCount = list(
    floorplanCount = 123,
    interiorCountPerFloorplan = 123
  ),
  tags = list(
    "string"
  ),
  worldTags = list(
    "string"
  )
)
}
}

\keyword{internal}
