% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_robot_application}
\alias{robomaker_delete_robot_application}
\title{Deletes a robot application}
\usage{
robomaker_delete_robot_application(application, applicationVersion)
}
\arguments{
\item{application}{[required] The Amazon Resource Name (ARN) of the the robot application.}

\item{applicationVersion}{The version of the robot application to delete.}
}
\value{
An empty list.
}
\description{
Deletes a robot application.
}
\section{Request syntax}{
\preformatted{svc$delete_robot_application(
  application = "string",
  applicationVersion = "string"
)
}
}

\keyword{internal}
