% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_simulation_application}
\alias{robomaker_delete_simulation_application}
\title{Deletes a simulation application}
\usage{
robomaker_delete_simulation_application(application, applicationVersion)
}
\arguments{
\item{application}{[required] The application information for the simulation application to delete.}

\item{applicationVersion}{The version of the simulation application to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a simulation application.
}
\section{Request syntax}{
\preformatted{svc$delete_simulation_application(
  application = "string",
  applicationVersion = "string"
)
}
}

\keyword{internal}
