% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_world_template}
\alias{robomaker_delete_world_template}
\title{Deletes a world template}
\usage{
robomaker_delete_world_template(template)
}
\arguments{
\item{template}{[required] The Amazon Resource Name (arn) of the world template you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes a world template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_world_template(
  template = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
