% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_simulation_application}
\alias{robomaker_describe_simulation_application}
\title{Describes a simulation application}
\usage{
robomaker_describe_simulation_application(application,
  applicationVersion)
}
\arguments{
\item{application}{[required] The application information for the simulation application.}

\item{applicationVersion}{The version of the simulation application to describe.}
}
\description{
Describes a simulation application.
}
\section{Request syntax}{
\preformatted{svc$describe_simulation_application(
  application = "string",
  applicationVersion = "string"
)
}
}

\keyword{internal}
