% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_simulation_job_batch}
\alias{robomaker_describe_simulation_job_batch}
\title{Describes a simulation job batch}
\usage{
robomaker_describe_simulation_job_batch(batch)
}
\arguments{
\item{batch}{[required] The id of the batch to describe.}
}
\description{
Describes a simulation job batch.
}
\section{Request syntax}{
\preformatted{svc$describe_simulation_job_batch(
  batch = "string"
)
}
}

\keyword{internal}
