% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world}
\alias{robomaker_describe_world}
\title{Describes a world}
\usage{
robomaker_describe_world(world)
}
\arguments{
\item{world}{[required] The Amazon Resource Name (arn) of the world you want to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  generationJob = "string",
  template = "string",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a world.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_world(
  world = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
