% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world_generation_job}
\alias{robomaker_describe_world_generation_job}
\title{Describes a world generation job}
\usage{
robomaker_describe_world_generation_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (arn) of the world generation job to describe.}
}
\description{
Describes a world generation job.
}
\section{Request syntax}{
\preformatted{svc$describe_world_generation_job(
  job = "string"
)
}
}

\keyword{internal}
