% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world_template}
\alias{robomaker_describe_world_template}
\title{Describes a world template}
\usage{
robomaker_describe_world_template(template)
}
\arguments{
\item{template}{[required] The Amazon Resource Name (arn) of the world template you want to
describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  clientRequestToken = "string",
  name = "string",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  lastUpdatedAt = as.POSIXct(
    "2015-01-01"
  ),
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a world template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_world_template(
  template = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
