% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_get_world_template_body}
\alias{robomaker_get_world_template_body}
\title{Gets the world template body}
\usage{
robomaker_get_world_template_body(template, generationJob)
}
\arguments{
\item{template}{The Amazon Resource Name (arn) of the world template.}

\item{generationJob}{The Amazon Resource Name (arn) of the world generator job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  templateBody = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the world template body.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_world_template_body(
  template = "string",
  generationJob = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
