% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_robots}
\alias{robomaker_list_robots}
\title{Returns a list of robots}
\usage{
robomaker_list_robots(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{The \code{nextToken} value returned from a previous paginated \code{ListRobots}
request where \code{maxResults} was used and the results exceeded the value
of that parameter. Pagination continues from the end of the previous
results that returned the \code{nextToken} value.}

\item{maxResults}{When this parameter is used, \code{ListRobots} only returns \code{maxResults}
results in a single page along with a \code{nextToken} response element. The
remaining results of the initial request can be seen by sending another
\code{ListRobots} request with the returned \code{nextToken} value. This value can
be between 1 and 200. If this parameter is not used, then \code{ListRobots}
returns up to 200 results and a \code{nextToken} value if applicable.}

\item{filters}{Optional filters to limit results.

The filter names \code{status} and \code{fleetName} are supported. When filtering,
you must use the complete value of the filtered item. You can use up to
three filters, but they must be for the same named item. For example, if
you are looking for items with the status \code{Registered} or the status
\code{Available}.}
}
\description{
Returns a list of robots. You can optionally provide filters to retrieve
specific robots.
}
\section{Request syntax}{
\preformatted{svc$list_robots(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
