% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_simulation_job_batches}
\alias{robomaker_list_simulation_job_batches}
\title{Returns a list simulation job batches}
\usage{
robomaker_list_simulation_job_batches(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{ListSimulationJobBatches} request where \code{maxResults} was used and the
results exceeded the value of that parameter. Pagination continues from
the end of the previous results that returned the \code{nextToken} value.}

\item{maxResults}{When this parameter is used, \code{ListSimulationJobBatches} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{ListSimulationJobBatches} request with the returned
\code{nextToken} value.}

\item{filters}{Optional filters to limit results.}
}
\description{
Returns a list simulation job batches. You can optionally provide
filters to retrieve specific simulation batch jobs.
}
\section{Request syntax}{
\preformatted{svc$list_simulation_job_batches(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
