% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_simulation_jobs}
\alias{robomaker_list_simulation_jobs}
\title{Returns a list of simulation jobs}
\usage{
robomaker_list_simulation_jobs(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{\link[=robomaker_list_simulation_jobs]{list_simulation_jobs}} again and
assign that token to the request object's \code{nextToken} parameter. If
there are no remaining results, the previous response object's NextToken
parameter is set to null.}

\item{maxResults}{When this parameter is used,
\code{\link[=robomaker_list_simulation_jobs]{list_simulation_jobs}} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{\link[=robomaker_list_simulation_jobs]{list_simulation_jobs}}
request with the returned \code{nextToken} value. This value can be between 1
and 1000. If this parameter is not used, then
\code{\link[=robomaker_list_simulation_jobs]{list_simulation_jobs}} returns up to
1000 results and a \code{nextToken} value if applicable.}

\item{filters}{Optional filters to limit results.

The filter names \code{status} and \code{simulationApplicationName} and
\code{robotApplicationName} are supported. When filtering, you must use the
complete value of the filtered item. You can use up to three filters,
but they must be for the same named item. For example, if you are
looking for items with the status \code{Preparing} or the status \code{Running}.}
}
\value{
A list with the following syntax:\preformatted{list(
  simulationJobSummaries = list(
    list(
      arn = "string",
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      name = "string",
      status = "Pending"|"Preparing"|"Running"|"Restarting"|"Completed"|"Failed"|"RunningFailed"|"Terminating"|"Terminated"|"Canceled",
      simulationApplicationNames = list(
        "string"
      ),
      robotApplicationNames = list(
        "string"
      ),
      dataSourceNames = list(
        "string"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Returns a list of simulation jobs. You can optionally provide filters to
retrieve specific simulation jobs.
}
\section{Request syntax}{
\preformatted{svc$list_simulation_jobs(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
