% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_tags_for_resource}
\alias{robomaker_list_tags_for_resource}
\title{Lists all tags on a AWS RoboMaker resource}
\usage{
robomaker_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The AWS RoboMaker Amazon Resource Name (ARN) with tags to be listed.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists all tags on a AWS RoboMaker resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
