% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_world_generation_jobs}
\alias{robomaker_list_world_generation_jobs}
\title{Lists world generator jobs}
\usage{
robomaker_list_world_generation_jobs(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{ListWorldGenerationJobsRequest} again and assign that token to the
request object's \code{nextToken} parameter. If there are no remaining
results, the previous response object's NextToken parameter is set to
null.}

\item{maxResults}{When this parameter is used, \code{ListWorldGeneratorJobs} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{ListWorldGeneratorJobs} request with the returned
\code{nextToken} value. This value can be between 1 and 100. If this
parameter is not used, then \code{ListWorldGeneratorJobs} returns up to 100
results and a \code{nextToken} value if applicable.}

\item{filters}{Optional filters to limit results. You can use \code{status} and
\code{templateId}.}
}
\value{
A list with the following syntax:\preformatted{list(
  worldGenerationJobSummaries = list(
    list(
      arn = "string",
      template = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      status = "Pending"|"Running"|"Completed"|"Failed"|"PartialFailed"|"Canceling"|"Canceled",
      worldCount = list(
        floorplanCount = 123,
        interiorCountPerFloorplan = 123
      ),
      succeededWorldCount = 123,
      failedWorldCount = 123
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists world generator jobs.
}
\section{Request syntax}{
\preformatted{svc$list_world_generation_jobs(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
