% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_worlds}
\alias{robomaker_list_worlds}
\title{Lists worlds}
\usage{
robomaker_list_worlds(nextToken, maxResults, filters)
}
\arguments{
\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call
\code{\link[=robomaker_list_worlds]{list_worlds}} again and assign that token to
the request object's \code{nextToken} parameter. If there are no remaining
results, the previous response object's NextToken parameter is set to
null.}

\item{maxResults}{When this parameter is used, \code{\link[=robomaker_list_worlds]{list_worlds}} only
returns \code{maxResults} results in a single page along with a \code{nextToken}
response element. The remaining results of the initial request can be
seen by sending another \code{\link[=robomaker_list_worlds]{list_worlds}} request
with the returned \code{nextToken} value. This value can be between 1 and
100. If this parameter is not used, then
\code{\link[=robomaker_list_worlds]{list_worlds}} returns up to 100 results and a
\code{nextToken} value if applicable.}

\item{filters}{Optional filters to limit results. You can use \code{status}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  worldSummaries = list(
    list(
      arn = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      generationJob = "string",
      template = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists worlds.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_worlds(
  nextToken = "string",
  maxResults = 123,
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
