% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_register_robot}
\alias{robomaker_register_robot}
\title{Registers a robot with a fleet}
\usage{
robomaker_register_robot(fleet, robot)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  fleet = "string",
  robot = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Registers a robot with a fleet.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_robot(
  fleet = "string",
  robot = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
