% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_restart_simulation_job}
\alias{robomaker_restart_simulation_job}
\title{Restarts a running simulation job}
\usage{
robomaker_restart_simulation_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (ARN) of the simulation job.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Restarts a running simulation job.
}
\section{Request syntax}{
\preformatted{svc$restart_simulation_job(
  job = "string"
)
}
}

\keyword{internal}
