% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_tag_resource}
\alias{robomaker_tag_resource}
\title{Adds or edits tags for a AWS RoboMaker resource}
\usage{
robomaker_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the AWS RoboMaker resource you are
tagging.}

\item{tags}{[required] A map that contains tag keys and tag values that are attached to the
resource.}
}
\value{
An empty list.
}
\description{
Adds or edits tags for a AWS RoboMaker resource.

Each tag consists of a tag key and a tag value. Tag keys and tag values
are both required, but tag values can be empty strings.

For information about the rules that apply to tag keys and tag values,
see \href{https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/allocation-tag-restrictions.html}{User-Defined Tag Restrictions}
in the \emph{AWS Billing and Cost Management User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
