% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_untag_resource}
\alias{robomaker_untag_resource}
\title{Removes the specified tags from the specified AWS RoboMaker resource}
\usage{
robomaker_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the AWS RoboMaker resource you are
removing tags.}

\item{tagKeys}{[required] A map that contains tag keys and tag values that will be unattached from
the resource.}
}
\description{
Removes the specified tags from the specified AWS RoboMaker resource.

To remove a tag, specify the tag key. To change the tag value of an
existing tag key, use
\href{https://docs.aws.amazon.com/robomaker/latest/dg/API_TagResource.html}{\code{TagResource}}
.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
