% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_update_world_template}
\alias{robomaker_update_world_template}
\title{Updates a world template}
\usage{
robomaker_update_world_template(template, name, templateBody,
  templateLocation)
}
\arguments{
\item{template}{[required] The Amazon Resource Name (arn) of the world template to update.}

\item{name}{The name of the template.}

\item{templateBody}{The world template body.}

\item{templateLocation}{The location of the world template.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  name = "string",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  lastUpdatedAt = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates a world template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_world_template(
  template = "string",
  name = "string",
  templateBody = "string",
  templateLocation = list(
    s3Bucket = "string",
    s3Key = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
