% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_batch_delete_worlds}
\alias{robomaker_batch_delete_worlds}
\title{Deletes one or more worlds in a batch operation}
\usage{
robomaker_batch_delete_worlds(worlds)
}
\arguments{
\item{worlds}{[required] A list of Amazon Resource Names (arns) that correspond to worlds to
delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  unprocessedWorlds = list(
    "string"
  )
)
}
}
\description{
Deletes one or more worlds in a batch operation.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_worlds(
  worlds = list(
    "string"
  )
)
}
}

\keyword{internal}
