% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_world_export_job}
\alias{robomaker_create_world_export_job}
\title{Creates a world export job}
\usage{
robomaker_create_world_export_job(clientRequestToken, worlds,
  outputLocation, iamRole, tags)
}
\arguments{
\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{worlds}{[required] A list of Amazon Resource Names (arns) that correspond to worlds to
export.}

\item{outputLocation}{[required]}

\item{iamRole}{[required] The IAM role that the world export process uses to access the Amazon S3
bucket and put the export.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
world export job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  status = "Pending"|"Running"|"Completed"|"Failed"|"Canceling"|"Canceled",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  failureCode = "InternalServiceError"|"LimitExceeded"|"ResourceNotFound"|"RequestThrottled"|"InvalidInput"|"AccessDenied",
  clientRequestToken = "string",
  outputLocation = list(
    s3Bucket = "string",
    s3Prefix = "string"
  ),
  iamRole = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a world export job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_world_export_job(
  clientRequestToken = "string",
  worlds = list(
    "string"
  ),
  outputLocation = list(
    s3Bucket = "string",
    s3Prefix = "string"
  ),
  iamRole = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
