% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_world_template}
\alias{robomaker_create_world_template}
\title{Creates a world template}
\usage{
robomaker_create_world_template(clientRequestToken, name, templateBody,
  templateLocation, tags)
}
\arguments{
\item{clientRequestToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{name}{The name of the world template.}

\item{templateBody}{The world template body.}

\item{templateLocation}{The location of the world template.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
world template.}
}
\value{
A list with the following syntax:\preformatted{list(
  arn = "string",
  clientRequestToken = "string",
  createdAt = as.POSIXct(
    "2015-01-01"
  ),
  name = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Creates a world template.
}
\section{Request syntax}{
\preformatted{svc$create_world_template(
  clientRequestToken = "string",
  name = "string",
  templateBody = "string",
  templateLocation = list(
    s3Bucket = "string",
    s3Key = "string"
  ),
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
