% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_fleet}
\alias{robomaker_delete_fleet}
\title{Deletes a fleet}
\usage{
robomaker_delete_fleet(fleet)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a fleet.
}
\section{Request syntax}{
\preformatted{svc$delete_fleet(
  fleet = "string"
)
}
}

\keyword{internal}
