% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_robot}
\alias{robomaker_delete_robot}
\title{Deletes a robot}
\usage{
robomaker_delete_robot(robot)
}
\arguments{
\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\value{
An empty list.
}
\description{
Deletes a robot.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_robot(
  robot = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
