% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_delete_world_template}
\alias{robomaker_delete_world_template}
\title{Deletes a world template}
\usage{
robomaker_delete_world_template(template)
}
\arguments{
\item{template}{[required] The Amazon Resource Name (arn) of the world template you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a world template.
}
\section{Request syntax}{
\preformatted{svc$delete_world_template(
  template = "string"
)
}
}

\keyword{internal}
