% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_simulation_application}
\alias{robomaker_describe_simulation_application}
\title{Describes a simulation application}
\usage{
robomaker_describe_simulation_application(application,
  applicationVersion)
}
\arguments{
\item{application}{[required] The application information for the simulation application.}

\item{applicationVersion}{The version of the simulation application to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  arn = "string",
  name = "string",
  version = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      etag = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  simulationSoftwareSuite = list(
    name = "Gazebo"|"RosbagPlay",
    version = "string"
  ),
  robotSoftwareSuite = list(
    name = "ROS"|"ROS2",
    version = "Kinetic"|"Melodic"|"Dashing"
  ),
  renderingEngine = list(
    name = "OGRE",
    version = "string"
  ),
  revisionId = "string",
  lastUpdatedAt = as.POSIXct(
    "2015-01-01"
  ),
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a simulation application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_simulation_application(
  application = "string",
  applicationVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
