% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_describe_world_export_job}
\alias{robomaker_describe_world_export_job}
\title{Describes a world export job}
\usage{
robomaker_describe_world_export_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (arn) of the world export job to describe.}
}
\description{
Describes a world export job.
}
\section{Request syntax}{
\preformatted{svc$describe_world_export_job(
  job = "string"
)
}
}

\keyword{internal}
