% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_get_world_template_body}
\alias{robomaker_get_world_template_body}
\title{Gets the world template body}
\usage{
robomaker_get_world_template_body(template, generationJob)
}
\arguments{
\item{template}{The Amazon Resource Name (arn) of the world template.}

\item{generationJob}{The Amazon Resource Name (arn) of the world generator job.}
}
\description{
Gets the world template body.
}
\section{Request syntax}{
\preformatted{svc$get_world_template_body(
  template = "string",
  generationJob = "string"
)
}
}

\keyword{internal}
