% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_list_world_templates}
\alias{robomaker_list_world_templates}
\title{Lists world templates}
\usage{
robomaker_list_world_templates(nextToken, maxResults)
}
\arguments{
\item{nextToken}{If the previous paginated request did not return all of the remaining
results, the response object's \code{nextToken} parameter value is set to a
token. To retrieve the next set of results, call \code{ListWorldTemplates}
again and assign that token to the request object's \code{nextToken}
parameter. If there are no remaining results, the previous response
object's NextToken parameter is set to null.}

\item{maxResults}{When this parameter is used, \code{ListWorldTemplates} only returns
\code{maxResults} results in a single page along with a \code{nextToken} response
element. The remaining results of the initial request can be seen by
sending another \code{ListWorldTemplates} request with the returned
\code{nextToken} value. This value can be between 1 and 100. If this
parameter is not used, then \code{ListWorldTemplates} returns up to 100
results and a \code{nextToken} value if applicable.}
}
\description{
Lists world templates.
}
\section{Request syntax}{
\preformatted{svc$list_world_templates(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
