% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_sync_deployment_job}
\alias{robomaker_sync_deployment_job}
\title{Syncrhonizes robots in a fleet to the latest deployment}
\usage{
robomaker_sync_deployment_job(clientRequestToken, fleet)
}
\arguments{
\item{clientRequestToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{fleet}{[required] The target fleet for the synchronization.}
}
\value{
A list with the following syntax:\preformatted{list(
  arn = "string",
  fleet = "string",
  status = "Pending"|"Preparing"|"InProgress"|"Failed"|"Succeeded"|"Canceled",
  deploymentConfig = list(
    concurrentDeploymentPercentage = 123,
    failureThresholdPercentage = 123,
    robotDeploymentTimeoutInSeconds = 123,
    downloadConditionFile = list(
      bucket = "string",
      key = "string",
      etag = "string"
    )
  ),
  deploymentApplicationConfigs = list(
    list(
      application = "string",
      applicationVersion = "string",
      launchConfig = list(
        packageName = "string",
        preLaunchFile = "string",
        launchFile = "string",
        postLaunchFile = "string",
        environmentVariables = list(
          "string"
        )
      )
    )
  ),
  failureReason = "string",
  failureCode = "ResourceNotFound"|"EnvironmentSetupError"|"EtagMismatch"|"FailureThresholdBreached"|"RobotDeploymentAborted"|"RobotDeploymentNoResponse"|"RobotAgentConnectionTimeout"|"GreengrassDeploymentFailed"|"InvalidGreengrassGroup"|"MissingRobotArchitecture"|"MissingRobotApplicationArchitecture"|"MissingRobotDeploymentResource"|"GreengrassGroupVersionDoesNotExist"|"LambdaDeleted"|"ExtractingBundleFailure"|"PreLaunchFileFailure"|"PostLaunchFileFailure"|"BadPermissionError"|"DownloadConditionFailed"|"BadLambdaAssociated"|"InternalServerError"|"RobotApplicationDoesNotExist"|"DeploymentFleetDoesNotExist"|"FleetDeploymentTimeout",
  createdAt = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Syncrhonizes robots in a fleet to the latest deployment. This is helpful
if robots were added after a deployment.
}
\section{Request syntax}{
\preformatted{svc$sync_deployment_job(
  clientRequestToken = "string",
  fleet = "string"
)
}
}

\keyword{internal}
